Consolidate and clean types
