+++ app/javascript/components/App.tsx
@@ -1,13 +1,8 @@
 import React, { useState, useEffect } from "react";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
-import { Deliveries, Home, Groceries, LanguageSelector, MasterNavbar, Order, Users } from "../components";
+import { Deliveries, Home, Groceries, LanguageSelector, MasterNavbar, Order, OrdersIndex, Users } from "../components";
 import { UserProvider } from './UserContext';
-
-interface FetchUserResponse {
-    id: string;
-    name: string;
-    email: string;
-}
+import { FetchUserResponse } from './api-types';
 
 const App = (): JSX.Element => {
     const [state, setState] = useState('');
@@ -41,6 +36,7 @@ const App = (): JSX.Element => {
                     <Route path="/" exact component={Home} />
                     <Route path="/groceries" exact component={Groceries} />
                     <Route path="/orders/:id" exact component={Order} />
+                    <Route path="/orders" exact component={OrdersIndex} />
                     <Route path="/deliveries" exact component={Deliveries} />
                     <Route path="/users" exact component={Users} />
                 </Switch>
